Adjust the overall dictionary architecture
